requirement.txt
aniso8601==7.0.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
Authlib==1.3.2
azure-core==1.33.0
azure-cosmos==4.9.0
azure-identity==1.20.0
azure-storage-blob==12.24.1
backcall==0.2.0
beautifulsoup4==4.13.3
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
contourpy==1.1.1
cryptography==44.0.1
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.14
decorator==5.2.1
dnspython==2.6.1
ecdsa==0.19.1
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-GraphQL==2.0.1
flask-swagger-ui==4.11.1
fonttools==4.56.0
fpdf==1.7.2
gdown==5.2.0
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
htmlmin==0.1.12
httplib2==0.22.0
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
msal==1.31.1
msal-extensions==1.2.0
multimethod==1.10
nest-asyncio==1.6.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
phik==0.12.4
pickleshare==0.7.5
pillow==10.4.0
pkg_resources==0.0.0
platformdirs==4.3.6
portalocker==2.10.1
promise==2.3
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.10.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.4.0
reportlab==3.6.12
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
Rx==1.6.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
statsmodels==0.14.1
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
visions==0.7.6
wcwidth==0.2.13
Werkzeug==3.0.6
wordcloud==1.9.4
ydata-profiling==4.12.0
zipp==3.20.2
